requirements.txt
mysql-connector-python>=8.2.0
PyYAML>=6.0
openai>=1.12.0
sqlparse>=0.5.0
